% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readElectionZipFile}
\alias{readElectionZipFile}
\title{readElectionZipFile}
\usage{
readElectionZipFile(zipfile)
}
\arguments{
\item{file}{The zip file from valmyndigheten. (Här kan vi lägga till en internetadress)}
}
\value{
A ValXML class object.
}
\description{
Read in the zipfiles from valmyndigheten to an R-object to parse.
}

